move schema-injection-utils to internal, was public in v1 for connecting to custom azure fluid relay instance, AzureLiveShareClient can be used instead
