+++ packages/live-share/src/index.ts
@@ -3,17 +3,23 @@
  * Licensed under the Microsoft Live Share SDK License.
  */
 
+export * from "./AzureLiveShareClient";
+export * from "./AzureLiveShareHost";
+export * from "./errors";
 export * from "./HostTimestampProvider";
 export * from "./interfaces";
+export * from "./LiveDataObject";
 export * from "./LiveEvent";
 export * from "./LiveEventScope";
 export * from "./LiveEventSource";
 export * from "./LiveEventTarget";
 export * from "./LiveEventTimer";
+export * from "./LiveFollowMode";
+export * from "./LiveObjectSynchronizer";
 export * from "./LivePresence";
 export * from "./LivePresenceUser";
-export * from "./LiveFollowMode";
 export * from "./LiveShareClient";
+export * from "./LiveShareRuntime";
 export * from "./LiveState";
 export * from "./LiveTelemetryLogger";
 export * from "./LiveTimer";
@@ -21,17 +27,11 @@ export * from "./LocalTimestampProvider";
 export * from "./TestLiveShareHost";
 export * from "./TimeInterval";
 export * from "./TimestampProvider";
-export * from "./schema-injection-utils";
-export * from "./LiveDataObject";
-export * from "./LiveShareRuntime";
-export * from "./AzureLiveShareClient";
-export * from "./AzureLiveShareHost";
-export * from "./LiveObjectSynchronizer";
-export * from "./errors";
 
 // TODO: our package structure is such that things that need to be exposed at all for other live share libs
 // are in the root of the directory even if they are "internal". move some to internal again?
+export * from "./internals/BaseLiveShareClient";
 export * from "./internals/DynamicObjectRegistry";
 export * from "./internals/DynamicObjectManager";
-export * from "./internals/BaseLiveShareClient";
+export * from "./internals/schema-injection-utils";
 export * from "./internals/smuggle";
